+++ share/time/timesup.qml
@@ -135,6 +135,7 @@ PanelWindow {
                     }
 
                     Text {
+                        font.family: theme.fontFamily
                         anchors.horizontalCenter: parent.horizontalCenter
                         text: "Time's up, " + root.displayName + "!"
                         color: theme.foreground
@@ -143,6 +144,7 @@ PanelWindow {
                     }
 
                     Text {
+                        font.family: theme.fontFamily
                         width: parent.width
                         text: "Your screen time for today is done."
                         color: theme.caption
@@ -152,6 +154,7 @@ PanelWindow {
                     }
 
                     Text {
+                        font.family: theme.fontFamily
                         anchors.horizontalCenter: parent.horizontalCenter
                         text: "Finishing in " + root.secondsLeft + "s"
                         color: root.secondsLeft <= 10 ? theme.error : theme.warning
@@ -176,6 +179,7 @@ PanelWindow {
                                 width: parent.width - 16
                                 spacing: 4
                                 Text {
+                                    font.family: theme.fontFamily
                                     width: parent.width
                                     text: "Ask a grown-up"
                                     color: theme.foreground
@@ -185,6 +189,7 @@ PanelWindow {
                                     wrapMode: Text.WordWrap
                                 }
                                 Text {
+                                    font.family: theme.fontFamily
                                     width: parent.width
                                     text: "for more time"
                                     color: theme.caption